server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Vladimir Moshchuk <ardonplay@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.16"
fastapi = "^0.110.2"
sentence-transformers = "^2.7.0"
uvicorn = "^0.29.0"
accelerate = "^0.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
